refining loading view
