+++ sphinx/Scenes/Badges/Member/Models/MemberBadgeDetailVM.swift
@@ -29,6 +29,7 @@ class MemberBadgeDetailVM : NSObject {
             if(isLoading){
                 let loadingViewFrame = vc.detailView.frame
                 vc.loadingView = UIView(frame: loadingViewFrame)
+                vc.loadingView?.isUserInteractionEnabled = false
                 vc.loadingView?.backgroundColor = UIColor.Sphinx.Body
                 if let loadingView = vc.loadingView{
                     let imageView : UIImageView = UIImageView(frame: loadingView.frame)
@@ -128,7 +129,7 @@ class MemberBadgeDetailVM : NSObject {
     
     func loadProfileData() {
         guard let person = message?.person else {
-            //dismissView()
+            vc.dismiss(animated: true)
             return
         }
         
@@ -204,43 +205,3 @@ extension MemberBadgeDetailVM : UITableViewDelegate,UITableViewDataSource{
 
 
 
-class ShimmerView : UIView {
-
-    var gradientColorOne : CGColor = UIColor(white: 0.85, alpha: 1.0).cgColor
-    var gradientColorTwo : CGColor = UIColor(white: 0.95, alpha: 1.0).cgColor
-    
-    
-    
-    func addGradientLayer() -> CAGradientLayer {
-        
-        let gradientLayer = CAGradientLayer()
-        
-        gradientLayer.frame = self.bounds
-        gradientLayer.startPoint = CGPoint(x: 0.0, y: 1.0)
-        gradientLayer.endPoint = CGPoint(x: 1.0, y: 1.0)
-        gradientLayer.colors = [gradientColorOne, gradientColorTwo, gradientColorOne]
-        gradientLayer.locations = [0.0, 0.5, 1.0]
-        self.layer.addSublayer(gradientLayer)
-        
-        return gradientLayer
-    }
-    
-    func addAnimation() -> CABasicAnimation {
-       
-        let animation = CABasicAnimation(keyPath: "locations")
-        animation.fromValue = [-1.0, -0.5, 0.0]
-        animation.toValue = [1.0, 1.5, 2.0]
-        animation.repeatCount = .infinity
-        animation.duration = 0.9
-        return animation
-    }
-    
-    func startShimmerAnimation() {
-        
-        let gradientLayer = addGradientLayer()
-        let animation = addAnimation()
-       
-        gradientLayer.add(animation, forKey: animation.keyPath)
-    }
-
-}
